+++ Controls/_popup/Opener/BaseOpenerUtil.ts
@@ -1,6 +1,4 @@
 import {Control} from 'UI/Base';
-import {parse as parserLib, load} from 'Core/library';
-import * as isEmpty from 'Core/helpers/Object/isEmpty';
 import ManagerController from 'Controls/_popup/Manager/ManagerController';
 import * as isNewEnvironment from 'Core/helpers/isNewEnvironment';
 import {goUpByControlTree} from 'UI/Focus';
@@ -19,51 +17,6 @@ let ManagerWrapperCreatingPromise; // TODO: Compatible
 let isLayerCompatibleLoaded; // TODO: Compatible
 
 export default {
-    requireModule(module: string|Control): Promise<Control> {
-        const moduleClass = this.getModule(module);
-        if (moduleClass) {
-            return Promise.resolve(moduleClass);
-        }
-        return load(module);
-    },
-
-    // _prepareModuleName(module: string): IModuleInfo {
-    //     const parsedModule = parserLib(module);
-    //     const isDefined = require.defined(parsedModule.name);
-    //     const moduleClass = isDefined && require(parsedModule.name);
-    //     return {
-    //         parsedModule,
-    //         isDefined,
-    //         moduleClass
-    //     };
-    // },
-
-    getModule(module: string|Control): Control|null {
-        if (typeof module === 'string') {
-            const parsedModule = parserLib(module);
-            const isDefined = require.defined(parsedModule.name);
-            let moduleClass = isDefined && require(parsedModule.name);
-            // Если кто-то позвал загрузку модуля, но она еще не отстрелила, require может вернуть пустой объект
-            if (!isDefined || isEmpty(moduleClass)) {
-                return null;
-            }
-            if (parsedModule.path.length) {
-                parsedModule.path.forEach((property) => {
-                    if (moduleClass && typeof moduleClass === 'object' && property in moduleClass) {
-                        moduleClass = moduleClass[property];
-                    }
-                });
-            }
-
-            // It's not a library notation so mind the default export for ES6 modules
-            if (moduleClass && moduleClass.__esModule && moduleClass.default) {
-                moduleClass = moduleClass.default;
-            }
-            return moduleClass;
-        }
-        return module;
-    },
-
     loadCompatibleLayer(callback: Function): void {
         const layerCompatibleModuleName: string = 'Lib/Control/LayerCompatible/LayerCompatible';
         const loadedCallback = () => {
